feat(로또): 로또를 통한 총 수익률을 구하는 기능 구현

+++ src/main/java/lotto/domain/buyer/Buyer.java
@@ -4,5 +4,5 @@ public interface Buyer {
 
     int enterToBuyLottoForMoney();
 
-    long calculateProfitRatio(int amount);
+    double calculateProfitRatio(int amount);
 }
